+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -22,6 +22,7 @@
 #### Other Changes
 * Disabled initialization of client telemetry background threads if client telemetry is disabled - See [PR 34889](https://github.com/Azure/azure-sdk-for-java/pull/34889)
 * Removed synchronized locking on generating random UUIDs - See [PR 34879](https://github.com/Azure/azure-sdk-for-java/pull/34879)
+* Capture diagnostics for cancelled `RNTBD` requests - See [PR 34912](https://github.com/Azure/azure-sdk-for-java/pull/34912)
 
 ### 4.44.0 (2023-04-21)